+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.menu.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
