+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,7 +1,7 @@
 package kitchenpos.product.application;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.order.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
